requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-api-python-client==2.159.0
google-auth==2.38.0
python-dotenv==1.0.1
httpx==0.28.1
tzdata==2025.1

# Testing
pytest==8.3.5
